+++ src/world/chunkManager.c
@@ -101,8 +101,6 @@ static Chunk_t **chunks_getChunks(State_t *restrict pSTATE, const Vec3i_t *restr
     return ppChunks;
 }
 
-/// @brief Returns an array (length = CUBE_FACE_COUNT) of pointers to each loaded chunk or null if it is not loaded around
-/// neighbor chunk at CHUNK_POS. (Heap)
 Chunk_t **chunkManager_getChunkNeighbors(const State_t *pSTATE, const Vec3i_t CHUNK_POS)
 {
     Chunk_t **ppChunks = malloc(sizeof(Chunk_t *) * CUBE_FACE_COUNT);
@@ -115,208 +113,12 @@ Chunk_t **chunkManager_getChunkNeighbors(const State_t *pSTATE, const Vec3i_t CH
         int ny = CHUNK_POS.y + pCMATH_CUBE_NEIGHBOR_OFFSETS[cubeFace].y;
         int nz = CHUNK_POS.z + pCMATH_CUBE_NEIGHBOR_OFFSETS[cubeFace].z;
 
-        // logs_log(LOG_DEBUG, "Checking neighbor of chunk (%d, %d, %d) at (%d, %d, %d).",
-        //          CHUNK_POS.x, CHUNK_POS.y, CHUNK_POS.z,
-        //          nx, ny, nz);
-
         Vec3i_t nChunkPos = {nx, ny, nz};
         ppChunks[cubeFace] = chunkManager_getChunk(pSTATE, nChunkPos);
     }
 
     return ppChunks;
 }
-#pragma endregion
-#pragma region Create Mesh
-static const uint8_t pCCW_QUAD_VERTS[6] = {0, 1, 3, 0, 3, 2};
-
-static inline void write_face_indices_u32(uint32_t *pIndicies, uint32_t base)
-{
-    pIndicies[0] = base + pCCW_QUAD_VERTS[0];
-    pIndicies[1] = base + pCCW_QUAD_VERTS[1];
-    pIndicies[2] = base + pCCW_QUAD_VERTS[2];
-    pIndicies[3] = base + pCCW_QUAD_VERTS[3];
-    pIndicies[4] = base + pCCW_QUAD_VERTS[4];
-    pIndicies[5] = base + pCCW_QUAD_VERTS[5];
-}
-
-static inline Vec3u8_t wrap_to_neighbor_local(uint8_t nx, uint8_t ny, uint8_t nz, CubeFace_t face)
-{
-    switch (face)
-    {
-    case CUBE_FACE_RIGHT: // nx == CHUNK_AXIS_LENGTH
-        return (Vec3u8_t){0, ny, nz};
-
-    case CUBE_FACE_LEFT: // nx == -1
-        return (Vec3u8_t){(uint8_t)(CHUNK_AXIS_LENGTH - 1), ny, nz};
-
-    case CUBE_FACE_TOP: // ny == CHUNK_AXIS_LENGTH
-        return (Vec3u8_t){nx, 0, nz};
-
-    case CUBE_FACE_BOTTOM: // ny == -1
-        return (Vec3u8_t){nx, (uint8_t)(CHUNK_AXIS_LENGTH - 1), nz};
-
-    case CUBE_FACE_FRONT: // nz == CHUNK_AXIS_LENGTH
-        return (Vec3u8_t){nx, ny, 0};
-
-    case CUBE_FACE_BACK: // nz == -1
-        return (Vec3u8_t){nx, ny, (uint8_t)(CHUNK_AXIS_LENGTH - 1)};
-    }
-
-    // This should never happen
-    return (Vec3u8_t){0, 0, 0};
-}
-
-bool chunk_mesh_create(State_t *pState, Chunk_t *pChunk)
-{
-    if (!pState || !pChunk || !pChunk->pBlockVoxels)
-        return false;
-
-    Chunk_t **ppNeighbors = chunkManager_getChunkNeighbors(pState, pChunk->chunkPos);
-    if (!ppNeighbors)
-        return false;
-
-    for (uint8_t i = 0; i < CUBE_FACE_COUNT; i++)
-    {
-        Chunk_t *pN = ppNeighbors[i];
-        // Chunks may not have loaded neighbors
-        if (pN && !pN->pTransparencyGrid)
-            return false;
-    }
-
-    Vec3u8_t *pNEIGHBOR_BLOCK_POS = cmath_chunk_blockNeighborPoints_Get();
-    bool *pNEIGHBOR_BLOCK_IN_CHUNK = cmath_chunk_blockNeighborPointsInChunkBool_Get();
-    if (!pNEIGHBOR_BLOCK_POS || !pNEIGHBOR_BLOCK_IN_CHUNK)
-        return false;
-
-    // max faces in a chunk possible (all transparent faces like glass)
-    const uint32_t MAX_VERTEX_COUNT = (uint32_t)CHUNK_BLOCK_CAPACITY * CUBE_FACE_COUNT * VERTS_PER_FACE;
-    const uint32_t MAX_INDEX_COUNT = (uint32_t)CHUNK_BLOCK_CAPACITY * CUBE_FACE_COUNT * INDICIES_PER_FACE;
-
-    ShaderVertexVoxel_t *pVertices = malloc(sizeof(ShaderVertexVoxel_t) * MAX_VERTEX_COUNT);
-    uint32_t *pIndices = malloc(sizeof(uint32_t) * MAX_INDEX_COUNT);
-    if (!pVertices || !pIndices)
-    {
-        free(pVertices);
-        free(pIndices);
-        free(ppNeighbors);
-        return false;
-    }
-
-    uint32_t vertexCursor = 0;
-    uint32_t indexCursor = 0;
-
-    const Vec3u8_t *pPOINTS = cmath_chunkPoints_Get();
-    if (!pPOINTS)
-        return false;
-
-    for (size_t i = 0; i < CMATH_CHUNK_POINTS_COUNT; i++)
-    {
-        const BlockDefinition_t *pBLOCK = pChunk->pBlockVoxels[i].pBLOCK_DEFINITION;
-
-        if (pBLOCK->BLOCK_ID == BLOCK_ID_AIR)
-            continue;
-
-#pragma region Neighbor Check
-        for (int face = 0; face < 6; ++face)
-        {
-            const Vec3u8_t N_POS = pNEIGHBOR_BLOCK_POS[cmath_blockNeighborIndex(i, face)];
-            const bool IN_CHUNK = pNEIGHBOR_BLOCK_IN_CHUNK[cmath_blockNeighborIndex(i, face)];
-
-            if (IN_CHUNK)
-            {
-                if (pChunk->pTransparencyGrid->pGrid[chunkSolidityGrid_index16(N_POS.x, N_POS.y, N_POS.z)] != SOLIDITY_TRANSPARENT)
-                    continue;
-            }
-            else
-            {
-                Chunk_t *pN = ppNeighbors[face];
-                if (pN)
-                {
-                    const Vec3u8_t POS_WRAP = wrap_to_neighbor_local(N_POS.x, N_POS.y, N_POS.z, (CubeFace_t)face);
-                    if (pN->pTransparencyGrid->pGrid[chunkSolidityGrid_index16(POS_WRAP.x, POS_WRAP.y, POS_WRAP.z)] != SOLIDITY_TRANSPARENT)
-                        continue;
-                }
-            }
-#pragma endregion
-#pragma region Face Creation
-            const FaceTexture_t TEX = pBLOCK->pFACE_TEXTURES[face];
-            const AtlasRegion_t *pATLAS_REGION = &pState->renderer.pAtlasRegions[TEX.atlasIndex];
-            const Vec3i_t BASE_POS = cmath_vec3u8_to_vec3i(pPOINTS[i]);
-
-            // Copy per-face vertices
-            for (int v = 0; v < VERTS_PER_FACE; ++v)
-            {
-                ShaderVertexVoxel_t vert = {0};
-                vert.pos = cmath_vec3i_to_vec3f(cmath_vec3i_add_vec3i(BASE_POS, pFACE_POSITIONS[face][v]));
-                vert.color = COLOR_WHITE;
-                vert.atlasIndex = TEX.atlasIndex;
-                vert.faceID = face;
-                pVertices[vertexCursor + v] = vert;
-            }
-
-            assignFaceUVs(pVertices, vertexCursor, pATLAS_REGION, TEX.rotation);
-
-            uint32_t base = vertexCursor;
-            write_face_indices_u32(&pIndices[indexCursor], base);
-
-            vertexCursor += VERTS_PER_FACE;
-            indexCursor += INDICIES_PER_FACE;
-        }
-    }
-
-    if (indexCursor == 0 || vertexCursor == 0)
-    {
-        free(pVertices);
-        free(pIndices);
-        free(ppNeighbors);
-
-        pChunk->pRenderChunk = NULL;
-        // Meshing succeeded and nothing to draw (full chunk and surrounded)
-        logs_log(LOG_DEBUG, "Render chunk is NULL for chunk at (%d, %d, %d) because it is entirely solid with loaded neighbors. \
-(Nothing to render).",
-                 pChunk->chunkPos.x, pChunk->chunkPos.y, pChunk->chunkPos.z);
-        return true;
-    }
-#pragma endregion
-#pragma region RenderChunk
-    // Shrink to used size <= max allocation
-    ShaderVertexVoxel_t *pFinalVerts = realloc(pVertices, sizeof(ShaderVertexVoxel_t) * vertexCursor);
-    uint32_t *pFinalIndices = realloc(pIndices, sizeof(uint32_t) * indexCursor);
-    if (!pFinalVerts)
-        pFinalVerts = pVertices;
-    if (!pFinalIndices)
-        pFinalIndices = pIndices;
-
-    vertexBuffer_createFromData_Voxel(pState, pFinalVerts, vertexCursor);
-    indexBuffer_createFromData(pState, pFinalIndices, indexCursor);
-
-    RenderChunk_t *pRenderChunk = malloc(sizeof(RenderChunk_t));
-    if (!pRenderChunk)
-    {
-        free(pFinalVerts);
-        free(pFinalIndices);
-        free(ppNeighbors);
-        return false;
-    }
-
-    pRenderChunk->vertexBuffer = pState->renderer.vertexBuffer;
-    pRenderChunk->vertexMemory = pState->renderer.vertexBufferMemory;
-    pRenderChunk->indexBuffer = pState->renderer.indexBuffer;
-    pRenderChunk->indexMemory = pState->renderer.indexBufferMemory;
-    pRenderChunk->indexCount = indexCursor;
-
-    free(pFinalVerts);
-    free(pFinalIndices);
-    free(ppNeighbors);
-
-    pChunk->pRenderChunk = pRenderChunk;
-
-    Vec3f_t worldPosition = cmath_vec3i_to_vec3f(chunkPos_to_worldOrigin(pChunk->chunkPos));
-    chunk_placeRenderInWorld(pChunk->pRenderChunk, &worldPosition);
-
-    return true;
-}
-#pragma endregion
 #pragma endregion
 #pragma region Allocate Voxels
 static Chunk_t *chunk_voxels_allocate(const State_t *pSTATE, const BlockDefinition_t *const *pBLOCK_DEFINITIONS, const Vec3i_t CHUNK_POS)
